Add basic support for dynamic dependencies
`trimja` cannot correctly trim a ninja file in which dynamic
dependencies are used.

Imagine a `build.ninja` file that compiles a set of C++ source files.
We do not know what header files are included by these sources until the
compiler compiles them and outputs the includes to ninja.  Ninja will
store these in the `.ninja_deps` database and read this for future
invocations to perform fast incremental builds.

If we asked `trimja` to restrict this build file to only those build
commands that depend on a `header.h` file, it would be impossible to
know which of the C++ source files depend on this until we compile.

This commit allows `trimja` to read the `.ninja_deps` file (though the
restriction is that it has to be in the same directory as the
`build.ninja` file, but we should support the `builddir` variable at
some point) and use this to build the graph.

This changes the plan on how `trimja` will be used to improve the time
taken to run CI on pull requests.  Now we will be required to store the
`.ninja_deps` file on the `main` branch and use this on pull requests.
This is easily done with the GitHub Action cache.

In order to write unit tests against the `.ninja_deps` file, a new
utility `kinja` has been written to create this binary file from JSON
and to do the reverse conversion.

3rd party libraries have been moved to `thirdparty`.