Add `trimja::ninja_clock`
Create a `std::chrono` clock that can represent the modification times
stored within `.ninja_deps`.  I realised that on little-endian machines
we can read/write a `std::uint64_t` directly, which simplifies things
enormously.

Now that we have our own `ninja_clock::time_point`, we can specialize
`std::chrono::clock_time_conversion` to convert between this and
`file_clock`.  This allows the reader/writer code to not know about this
conversion and instead put it in a single place outside this code.

As ninja has platform specific code when writing `.ninja_deps` we will
have to have the same and this is more easily done in `ninja_clock.cpp`.