Add Linux support
Add support for converting between `file_clock` and `ninja_clock` on
non-Windows platforms. This should be simpler than Windows as it looks
like it is Unix time in nanoseconds.

We require a minimum of Clang 17 in order to support
`clock_time_conversion` and printing/formatting
`std::chrono::file_time`.  This is why we choose the cutting edge
`ubuntu` version on the Github runner instead of `ubuntu-latest`.