Replace `quick_exit` with `_Exit`
MacOS does not support `std::quick_exit` but does support `_Exit`.  As
we are now no longer calling `std::at_quick_exit` there is not a need to
use `std::quick_exit` over `_Exit`.

Though we don't compile on MacOS right now it is a good idea to try and
minimize any differences when possible.