Add naive handling of `default`
Treat `default` like a `phony` edge that doesn't have a name.

This is not the ideal behaviour of `default` as I think the correct one
would only include those `in` paths that are required and remove (or
dummy out with a `phony` edge) those that aren't.

For example the right contents for `tests/trimja/default/expected.ninja`
should either be:

```
build a_out: phony a_in
build c_out: phony c_in
default a_out c_out
```

or

```
build a_out: phony a_in
build b_out: phony
build c_out: phony c_in
default a_out b_out c_out
```

I think we need a similar behaviour for `phony` build edges that are not
referred to by any other build edge since these are a bit like a
"named"-`default`.  These can probably be implemented in a similar way
later on.

The lexer for `default` doesn't quite trim trailing newlines in the same
way as it does for let/build statements so when running on Linux we have
LF in the output but CRLF in the expected.  To counter this we
change the expected files to end with just LF and then open the expected
file and the file we write to in binary mode.