Improve `default` semantics
For all paths specified in `default` that are not required by anything
else we don't really want to build them when running `ninja` without any
targets.

Instead of splitting up the `default` statement into different parts and
printing out only the required parts, we keep the `default` statement
identical and instead create empty `phony` build commands for output
statements that are not required but are mentioned.