Include all build outputs in trimmed file
Previously if we had a build command whose output was not needed we
would avoid printing it in the resulting ninja file.

After handling `default`, we needed a new state where we wanted to
include all inputs to `default`, but create `phony` rules for those
inputs if there were no other build command that required them.  This
allowed us not to have to edit the `default` statement and to not
bring in dependencies if they were only required by `default`.

If we didn't, then running `ninja` would build everything mentioned in
`default` every single time and wouldn't save us any time.  `default` is
just a nice way of grouping a set of outputs.

Similarly, we want to do the same with `phony` commands in the future.
So that something like `ninja all_tests` would build only the tests that
had changed.

Thinking harder on the subject, it is very possible that build
systems will have something like `ninja myexecutable.exe` that will
mention an output of a non-phony rule. In this case we want to build
it only if a dependent had changed.  But today we would not include that
build edge and `ninja` would report an error - probably not what we
want.

This commit changes `trimja` to output a build command for all
outputs mentioned in the original `ninja` build file, but create `phony`
edges for those not affected by the changed files. This way
`ninja myexecutable.exe` would return a success code and report that
there is no work to do.

In this commit we tidy up the string comparison code to string CR
characters instead of messing around with binary outputs.