Add tests for validations
After swapping to the behaviour where every output file gets at least a
`phony` command, we actually have a reasonable (although maybe not
perfect) support for validations.

This commit adds some unit tests and updates some internal
documentation.

The only behaviour today that is questionable, is when we have a build
command that does not depend on anything required, but it has a
validation edge that does depend on something that is required.  Right
now we would replace that build command with a `phony`, so that nothing
happens if it is run.  However, you could argue that validations should
universally be kept in `trimja` and then if that validation has no
required dependencies it will also be `phony`.

If `phony` build commands can have validation edges then this is
probably the behaviour we desire.  If they can't then it is probably
okay to not change the current behaviour since it is unlikely that a
build edge and its validations don't share all/most inputs.