Add initial `.ninja_log` support
The large remaining work in `trimja` is to look at `.ninja_log` to find
out whether build commands in the ninja build file are new or have
changed from the previous run and if so, unconditionally add them to the
trimmed file.

This is very basic support at the moment because we only hash the build
command after substituting in `$in` and `$out` (without dealing with
shell escaping) and not doing any other variable evaluation.

We update all of our unit tests to run `ninja` over the files first
(which has the added benefit of validating them) to generate the
`.ninja_deps` and `.ninja_log` files.  Then afterwards we can run
`trimja`, which will look at these files and hopefully decide that all
of the build commands are identical.  In the future we should have a
`bootstrap.ninja` file that is run by `ninja` so we can add, remove, or
modify build commands compared to the `build.ninja` file.

Use `ninja --version $in -> $out` in our tests because it is a program
that is going to be available for us, it will return very quickly
(`ninja` will print the version and ignore the rest), and it allows us
to encode the `$in` and `$out` variables in the build command.