Remove `kinja`
Remove `kinja` for the moment as my focus will be on getting `trimja`
working with historical `.ninja_log` and `.ninja_deps` files.

`kinja` was useful to print elements of `.ninja_deps` for testing, but
this code is now done and unlikely to change any time soon.  `kinja` was
also destined to mock out `.ninja_deps` for C++ based projects, but this
will be on the backburner for a while.