Swap CMake to ninja generator
When not using MSVC, use the ninja generator for CMake.  This should
give us a bit of a speed-up and we should really be using ninja stuff
as much as possible.

Using MSVC, CMake, and ninja together is difficult as it seems like we
would need to run the MSVC script to set up environment variables.  So
for now we just leave this configuration.