Add shell escaping in commands
When the `$in` and `$out` variables contain special characters, ninja
will add quotations and escape them.  As this affects the hash of the
build commands we need to take this into consideration.

While fixing this I realised that we don't escape the outputs in the
`phony` commands that we create because we don't escape spaces.  To fix
this this commit does a bunch of refactoring to store the partial
strings of the build command so that we don't need to encode the outputs
again, just print the string.

Include `CanonicalizePath` in `util.h` as we'll need this shortly to
normalize paths.