Add path normalization
ninja normalizes all paths (they call it canonicalization) and then the
normalized variables (`$out`, `$in`, `$in_newline`) are substituted into
the command.  As we use the command to generate our hash that compare to
the `.ninja_log` file - we should make sure we match what ninja
generates.  Otherwise we will assume that build commands not in the
canonical form are new and will never be `phony`ed out.