Support `include`
Support ninja build files `include`ing other ninja files.  This works
like C/C++ preprocessor `#include` directives where the processing is
identical to if the line was replaced with the file contents.

To keep things simple, we inline all `include`d files into the singular
output file.  If we didn't, then it becomes complex to decide on how we
output multiple files, e.g. would we need the user to specify the
`--outDir` in that case?

`subninja` has been explicitly marked as not supported yet.  It is
entirely possible to do so, but if we want to inline `subninja` files it
becomes a bit trickier as the files have different scopes.

In order to support `include` we have broken out many member variables
from `Parser` into `BuildContext`, which we can pass to another `Parser`
instance that is created for each `include`.

`Parser`, `BuildContext`, and friends should be moved to a new component
at a later point to reduce the amount of code inside `trimutil.cpp`.