Handle different path separators in Windows
On Windows, when ninja first encounters a path it becomes the canonical version
of all paths that differ only in path separators.  If we see
`foo/index.h` first in the ninja file, and then later encounter `foo\index.h`,
ninja sees these as equivalent and will use `foo/index.h` as the
canonical path, e.g.

```
rule echo
  command echo $out

build foo\index.h: echo
```

We will see "foo/index.h" printed out.

To do this we create a hash and equality struct for `unordered_map` that
ignores the path separator difference and we use that for all of our
path lookups.  The methods of `Graph` will now update the path
requested, changing it to the canonical one.