Update command line options
Rename `--restrict` to `--affected` as this is the name used by
some other build systems (e.g. nx and TurboRepo) to describe a way to
notify what files have been changed.

Require passing a single dash `-` when passing in the list of
affected files from stdin.  Previously I have had lots of issues by
accidentally not passing `--affected` and not piping in something to
stdin.  This leaves `trimja` waiting for input and gives the impression
that the application is hanging.  A singular dash is used by wget` and
recommended as an option by the "POSIX Utility Syntax Guidelines" for
tools that can take input on stdin or output to stdout.

Now passing neither `-` nor `--affected` will return quickly and let the
user know they have to pass at least one.