Fix `DepsReader` reading bug and improve error handling
`DepsReader` reads a binary file but was not passed an `std::ifstream`
in binary mode.  This commit changes this component to take the file
path in the constructor so that we guarantee we open the file in the
correct mode.  This also allows us to throw a nicer error message that
mentions the file name when things go wrong.

Use the `std::ifstream::exceptions` method instead of manually checking
the state and then throwing.