Add iterators to `DepsReader`
Follow the same design as `LogReader` and add `begin()` and `end()` to
`DepsReader`.  This allows us to use a range-for loop in `trimutil.cpp`
and keeps things looking tidy.

Update the iterator constructors to call `++*this;` instead of `read` as
that will reset themselves to compare equal to the sentinel value.

Move all implementations for `LogReader::iterator` to the source file.