Support `phony` and add `--explain` option
`phony` rules were not supported.  This commit adds them to the list of
rules in the constructor of `BuildContext` so it's understood by trimja.

However, `phony` and `default` rules are not the same as user-defined
rules. Only if we have a `phony` build edge that is depended on
(directly or transitively) by a user-defined build edge do we need to
include all of its edges.

The algorithm has been updated, simplified, and improved to handle
the new logic.

To help with debugging, an `--explain` argument has been added so that
we print to stderr whenever we decide to keep an edge and for what
reason.