Refactor parsing code
Decouple the parsing code from having to generate `BuildContext`.

Instead we have a parser that has an easier interface to give you all of
the relevant part and this can be used by `TrimUtil` to create
`BuildContext`.

`Parser` has been renamed to `ManifestParser` as we use the `Parser`
suffix for all of the classes in that component and ninja calls its
component `ManifestParser`.