Fix include and subninja paths
The paths specified by `include` and `subninja` statements are relative
to the current ninja build file and we weren't taking into account the
relative path between this ninja build file and our current working
directory.

This commit changes the snapshot tests to not run from the same
directory as the ninja build file and adds the relative path from the
current working directory on to new paths we need to follow.