Add `--builddir` flag
Add a command line flag to print out the path of the directory that
ninja will be outputting `.ninja_log` and `.ninja_deps`.  This will be
based on the `$builddir` variable as well as the relative path to the
ninja build file.

`--builddir` will be used by `trimja-action` to locate the build files
in order to add them to Github cache.

During this I discovered that the `consume` function doesn't work with
variable ranges since it never reads the name or value from the
`VariableReader`.  This was a bug that could affect trimming if there
was ever 2+ variables allowed underneath a `pool`.

`consume` has been moved to a `static` function so that it doesn't
conflict with the other free functions when doing a unity build.