Build on MacOS
Build on the lowest supported macos provided by the free Github Action
Runners so that we can run trimja on all those platforms without glibc
issues.

Give steps in the matrix a nicer name that doesn't continually change
when we change operating systems versions. This allows us to not have to
continually update the settings on what checks are required.