Support variables in subninja statements
Add partial support for `subninja` statements in a Ninja build file by
resetting variables that share a name when we exit the subninja scope.

Like `include`, `subninja` commands will be inlined into one Ninja file
by trimja.  In theory it would be much easier if we could rewrite all of
the individual files in-situ with `--write`, but that means we can't
support `subninja` statements writing to `stdout` or `--output`.  So
when exiting a file included with a `subninja` statement, we reset all
variables shared between parent and child that are now different to the
parent value.

A constructor for `BasicScope` has been manually written because
`fixed_string` doesn't have a copy constructor.  I am wary about adding
copy constructors to `fixed_string` as currently this catches situations
where we create accidental copies when find or inserting from a map.

The remaining part is to rename rules which may conflict with rules in a
parent scope.