Improve order-only dependency behaviour
For incremental builds, changes in order-only dependencies do not cause
a build command to be rerun by ninja.  The idea being that after their
first run, their real dependencies have been loaded into `.ninja_deps`
and, as they are more accurate, can be used instead.

trimja is trying to emulate an incremental build to rebuild only those
commands affected by a certain set of files.  So if an affected file is
mentioned as an order-only dependencies, but it doesn't appear for that
output in `.ninja_deps` - it wasn't **actually** needed.

This should reduce the number of build commands that we include in the
trimmed file.

Additionally, there was a bug where we incorrectly included the
order-only dependencies in the `$in` variables - `out5` in the
dependencies test case should cover this.