Use trimja-action
Use trimja ourselves to reduce the amount we need to build on pull
requests.  Unfortunately the Ninja file generated from CMake doesn't
have good dependencies for the tests.  The tests don't have a proper
dependency on the `trimja` executable, nor the test files that it reads.

For the moment we pass both `trimja` and `test` into `affected` so that
everything is always built.  However, we should look into improving
this.