Use trimja-action
Use trimja ourselves to reduce the amount we need to build on pull
requests.  Unfortunately the trimja executable is dependent on all code
files and all tests run from the executable so the only time we will
save is when we change something that isn't code (e.g. a README.md.  But
in this case we can skip the entire build.