Fix `default` generation with no affected inputs
When the `default` statement has no affected inputs we tries to `phony`
it out it fails because we don't set all of the `std::string_view`
pieces.  Instead, it is easier to force it to always be printed.

Treat `phony` rules the same, which doesn't change the output but saves
us some time regenerating the rule.