Print out all attempted paths
When trying to look up paths passed to `--affected` we attempt to
convert to absolute and relative paths in case the Ninja input file does
not match how files were passed in.

When this fails, it would be nice to let the user know that other file
paths were tried.  It is a good reminder what base path is used when
converting from relative to absolute paths.