Fix absolute/relative path handling in `--affected`
The conditional was incorrectly reversed due to my misreading of
`std::error_code` so we never attempted to look up the absolute or
relative path for paths passed in to `--affected`.

It was a bit tricky to do unit test for absolute paths as they need to
be generated beforehand.  This commit is less lazy and includes unit
tests for absolute and relative path handling by generating files in the
CMake generation step.