Swap to `boost::unordered_flat_map`
The performance of `std::unordered_map` is widely known to be
sub-optimal due to the requirement for iterator and reference stability.
This precludes

An up-to-date benchmark can be found here:
https://jacksonallan.github.io/c_cpp_hash_tables_benchmark, which shows
that `boost::unordered_flat_map` performs extremely well for inserts and
lookups, which is all we do.  The lack of iteration speed, the erasure
residual performance impact, and the minimum capacity of 30 all have no
affect on us.

The `boost::unordered_flat_map` implementation has been taken from
https://github.com/MikePopoloski/boost_unordered.

Add more implementation to `fixed_string` as we now need to be
move-constructible and move-assignable since we will be moved around
inside `boost::unordered_flat_map`.

Add `SYSTEM` to the include path for `thirdparty` so we don't get
warnings on external code.