Remove `fixed_string` temporaries
We can go back to using `is_transparent` and heterogenous lookup
since `boost::unordered_flat_map` supports it.  Previously we had to
remove it when supporting older compiler versions.

This removes a bunch of temporary copies (that probably could have been
avoided) and cleans up the code a lot.  The constructor taking a
`std::string_view` is marked as `explicit` so we do not accidentally
create `fixed_string`s if we pass a `std::string_view` into `find()` and
haven't used the right hash and equality predicate that support
`is_transparent`.