Improve parser interface
Make more methods `const` by doing work in the constructors of certain
`*Reader` classes.  This means that we don't have to worry as much about
calling `name()` before `value()` in the `VariableReader`, which is hard
to do if you are calling `foo(v.name(), v.value())` as there's no C++
guarantee which function is called first.

Add structured bindings to `VariableReader` to keep the interface even
more succinct.

Rename non-`const` methods to `read*()` so that it is more obvious that
they are mutating.