Skip destructors within `TrimUtil`
Based on benchmarks, running the destructor for all the objects in
`BuildContext` takes 1-2% of the total time.  Instead we can skip it
entirely by extending the lifetime of this object until we call
`std::_Exit` - ending the application and skipping destructing all
objects on the stack.

Using a `std::shared_ptr<void>` avoids us having to define
`BuildContext` in the header file and bring in a load more include
directives.  Another alternative is to make `TrimUtil` instantiable and
have `BuildContext` as a (perhaps pimple) member variable.  However,
this would require us to either reset all members in `BuildContext` when
we call `trim`, or to ban `trim` being called multiple times.  Neither
are as easy as the current chosen solution - even if it is a bit
strange.