Make `*Util` components constructible
Previously (603949f65421f5f3e055da36ac0888b4aa118267) we made some
changes to `TrimUtil` to return a `std::shared_ptr<void>` containing the
internal variables used in the algorithm.  This allowed us to defer
destruction in `TrimUtil` so that we can skip it with `std::_Exit` later
on - saving 1-2%.

The change did not sit well with me and is not very idiomatic.  This
commit changes both `BuildDirUtil` and `TrimUtil` to be constructible
and have their static methods changed to non-`const` member variables.
We store the variables in a `std::unique_ptr` to defer destruction of
the state until the destructor of the `*Util` - which is skipped when we
call `std::_Exit`.