Remove unnecessary `time_point` conversion
In general we don't need the `mtime` value from `DepsReader` for
anything in trimja.  So instead we change the member variable to
`ninja_clock::time_point` to avoid the conversion to
`std::chrono::file_clock::time_point` - even if it is cheap.

We could do the same thing that we did with `LogReader` and pass a
bitfield describing what fields we need.  However, it's extremely cheap
to read the fields since it's a bit copy.