Add clang-tidy support
Add a CMake option `ENABLE_CLANG_TIDY` to enable `clang-tidy` for the
project.

This is enabled on Linux only in the Github runners since it is the only
image that has `clang-tidy` installed by default.

I have disabled any check that didn't seem to improve the code.  The
only code that was exempt and should be fixed is the recursive function
calls in `trimutil.cpp` triggered by `misc-no-recursion`.  I think
fixing them should be done separately to avoid an overly complex commit.