Add AddressSanitizer support
Enable AddressSanitizer on all platforms in debug builds.  This will
slow down programs by about 2x so is quite acceptable in our use cases
where tests run in ~50ms.

Replace a `memcmp` with `strncmp` in `getopt.c` because it seems to not
compare byte-by-byte with an early exit on MSVC, and because of this is
accessible memory past the end of the string.  It is replaced with
`strncmp`, which fixes the issue.

Luckily we don't have any other issues arising in trimja that were
highlighted by AddressSanitizer.