Remove `resetValue` methods
libfuzzer has found an issue when we reassign the same variable multiple
times.  Given a Ninja build file,

```
var = cous
var = $var$var
```

The variable `var` should say "couscous" but our implementation would
first reset `var` to the empty string and then concatenate the empty
string with itself to give "".

However, libfuzzer also found a crash in this situation as we are
keeping references around and appending `std::string` to itself.  In the
case we need to reallocate we are have problems.

This commit removes `resetValue` and instead we create a new
`std::string` each time.  This is slower in the case where we overwrite
the same variable, however this is a rare situation and the common case
is to have new variables.

We never use the return value of the `set` methods so this has been
removed as well to avoid any similar issues in the future.