Fix `DepsReader` issues
libfuzzer has issues with `DepsReader` when the `.ninja_deps` is
malformed.  Mainly that we do not check whether the length of records is
sensible before doing arithmetic.  It was possible to supply a record
size of less than 4, which caused underflow and us to try and read a
huge number of bytes.

Additionally, the path record cannot have an empty path (according to
the Ninja code base) and the path must be padded to a multiple of 4
bytes.  So we need a minimum path length of 4 and checking for this
stops us decrementing the `std::string` iterator past the end.

In order to fuzz this component it was easier to pass a `std::istream&`
instead of `std::filesystem::path` since we needed to inject the corpus
instead of reason from a file.