Add fuzzing
Add 3 fuzzing entry points, each testing one of the 3 parsers we have:

  * `BuildDirUtil` (ninja build parser)
  * `DepsReader` (`.ninja_deps` parser)
  * `LogReader` (`.ninja_log` parser)

In order to do this we break up `trimja` project into a `core` static
library and the `trimja` executable.  This allows us to link the `core`
library into the separate fuzzing executables.

To keep things more managable, the `CMakeLists.txt` file has been neated
up, including moving some conditionals to generator expressions.  This
uncovered an issue where we weren't actually enabled Address Sanitizer
on MSVC because its generator generates multiple configurations at the
same time.  The sanitizers can't be enabled just yet because we need to
set up the path the correct DLL on Windows.  This can be done in a
follow up commit.

In an unrelated problem, `trimja` is failing when running on CI.  This
has been disabled for the moment until it can be fixed.