Conditionally add back LTO in debug mode
In the previous commit, link-time optimization was disabled for debug
mode when we enabled fuzzing.  This was disabled as I couldn't find a
way to enable both due to getting repeat link errors.  However, it is
an issue only with thin LTO and earlier versions of Clang.

This commit enables LTO (which for CMake is thin LTO) for later versions
of Clang that do not have this issue when fuzzing is enabled.