Add full testing for fuzzdepsreader
Update the `fuzzdepsreader` to roundtrip conversion from the ninja deps
binary format and check that the result is identical to the input.

In order to do this we have to tighten up the format that we expect and
start throwing on inputs that we don't expect.  Mainly this is checking
that the record size was a multiple of 4 and that we received the
correct padding bytes.

Run `fuzzdepsreader` for a few hours with `-max_len=1024` and commit all
the corpora generated.  Now we can use this to add an additional test
that runs all of these corpora and run this on CI.