Run fuzzing corpora on all platforms
Though fuzzing is not available in all platforms, we can at least run
the existing fuzzing corpora for all platforms to make sure that there
are no problems.  This is simple to do by editing our own entry point in
`nofuzzing.m.cpp` to iterate through files in a directory and running
these against the fuzzing entry point.