Add roundtrip fuzzing for `LogReader`
Add fuzzing for `LogReader` by roundtripping a corpus twice through
`LogWriter` and checking that the first and second outputs are
identical.

We have to roundtrip twice and not compare against the original since
there are numerous inputs that map to the same `LogEntry`, e.g. all
integers can have any number of leading zeros.  Perhaps in the future we
can tighten up the format so there is a 1-1 mapping, but this should be
fine for the moment.

`LogReader` is by no means optimal in terms of performance.  But there
is little reason to start optimizing now before proper benchmarks exist.