Prioritize most affected outputs
If a build is going to fail, we want it to fail as fast as possible.
This will give quicker feedback to the developer and reduce build
resources.

To facilitate this, we want to prioritize build edges that depend on
changed files over other build edges that are at the same dependency
level. e.g. for C++ projects we want to start compiling modified
source files before other sources files that belong to the same
library/executable.  In this case we shouldn't increase (on average) the
total build time on a successful fun, but an unsuccessful build would
fail much faster.

In ninja 1.12.0 a critical path schedule was added
https://github.com/ninja-build/ninja/pull/2177, which prioritizes build
commands based on their distance from the build target.  If two or more
build commands have the same distance, it looks like the one that
appears first in the build file is built first.

Before ninja 1.12.0, the one that appears first in the build file is
prioritized.

This commit floats all affected commands to the top of the ninja build
file (as far as it is possible at least) so that older ninja versions
will work, and it helps ninja 1.12.0 in the case of ties.

A future change will need to be made to artificially lengthen the
critical path for affected files so that ninja 1.12.0+ will be fully
supported.