Add `soagen` dependency to clean up code
In several places we use multiple arrays to store data about the nodes
in the graph.  These graphs always have the same size as the number of
nodes, but we keep them separate to achieve better cache locality.

`soagen` can do the same thing, but in a much neater way, while reducing
the number of allocations.