Bump Ubuntu runner from 20.04 to 22.04
The runner is now deprecated:

https://github.com/actions/runner-images/issues/11101

The only reason we are using the oldest version is so that all releases
will be built with a version of glibc that runs on all valid Github
Ubuntu runners.

We have to update to Clang 13 as Clang 12 is no longer in this image.
This brought up a few Clang Tidy issues for newly added rules.