Update Ubuntu version
Github Action Runner Ubuntu 20.04 is now deprecated so this change moves
to Ubuntu 22.04.

Unfortunately this added clang-tidy failures as more checks have been
added so these need to be fixed.  To avoid this happening again, we
build on the latest Ubuntu as well to catch some of these issues ahead
of time.