Extract `NestedScope`
Start taking out components from `trimutil.cpp` to simplify this
file, which is by far the most complex in the project.

`NestedScope` has been updated to not copy the latest `BasicScope`, but
instead to have a stack of these `BasicScope`s instead.  This should be
better for cases where the root ninja file has most of the variable
definitions.  However, subninja use is limited and real-world use cases
should be benchmarked if performance is ever an issue in this component.