Extract `PathVector` to `StringStack`
Start cleaning up components in `trimutil.cpp` to simplify this file
as it is the most complex component in `trimja`.

There is a slight change in behaviour, where we clear the `std::string`
in `emplace_back` right before we reuse it, instead of during `clear()`.
This is just a micro-optimization trying to avoid hitting a bunch of
cache-misses instead of just at most one when we clear the one string,
at a point where we are most likely to want to write to it very soon.