Improve debugging experience using `IndexInto`
Using raw indices to represent nodes in `Graph` has a huge benefit that
it's really easy to augment this data structure with additional
properties by creating a `std::vector<Foo>` with the same size as the
number of nodes in the graph - basically the ECS (entity-component
system) idiom.

The big downside is that debugging is particularly difficult since all
you have is a `std::size_t` and you have to manually look up values you
care about in other `std::vector`s to piece together `trimja`'s view of
the world.

This commit introduces `IndexInto`, effectively a `std::size_t` and a
pointer to the original container so that we can write useful natvis
files (and GDB pretty printers in the future if anyone needs this) that
displays all of the relevant data for that node.

In release mode (i.e. if `NDEBUG` is defined) then we no longer store
the pointer to the original container to keep the memory footprint as
lean as it was previously.

Future changes will introduce more uses of `IndexInto` inside
`trimutil.cpp` so that we can easily view extra information inside the
`isAffected` and `isRequired` `std::vector`s etc.

Add natvis files for `IndexIntoNoDebug`, `IndexIntoDebug`, and add the
natvis file for `gch::small_vector` from the original github repository
https://github.com/gharveymn/small_vector.  This is under the same MIT
license as mentioned in `thirdparts/gch/LICENSE`.