Add more `IndexInto` debugging help
Replace more `std::size_t` index values with `IndexInto` and appropriate
visualizations via natvis.

To support this we need to add a `TAG` template parameter to `IndexInto`
so that we can distinguish the different indices used within
`BuildContext`.  Also, since we have the `nodeToCommand` `std::vector`,
which is indexed by `Graph::Node`, it makes sense if the
`std::size_t m_index` member variable in `IndexInto` could be yet
another `IndexInto`, and we can see all of this information inside the
natvis files.

I had compilation issues using a sentinel value of `IndexInto` within
`tiny::optional` so I have swapped to `std::optional` for the moment,
and it will have to be something to be revisited in the future.  This is
the reason why `constexpr` was added to `IndexInto` and I think I'll
leave it there for the moment as it seems reasonable.

Implicit conversions are usually frowned upon in C++, but here I think
it makes sense to keep the code tidy, and we wouldn't have protected
against someone looking up an index in the wrong `vector` if it was
explicit or required a method call.

A future change should move the stack variables `isAffected` and
`isRequired`, which are indexed by `Graph::Node`, into `BuildContext` so
we can include these in the natvis information.