Make more things visible inside `BuildContext::Node`
Move `isRequired` and `isAffected` into `BuildContext` so that the
natvis for `IndexInto<BuildContext, NodeTag, Graph::Node>` can include
this information.

Swap uses of `Graph::Node` to `BuildContext::Node`, which is aliased to
the `IndexInto` just mentioned, throughout `trimutil.cpp` so that the
natvis visualization for these variables shows `isRequired` and
`isAffected`.