Introduce `pmr` instead of `forward_list`
Instead of storing `std::string` elements in our
`std::forward_list<std::string>` storage, instead we can "leak" them
into a `std::pmr::monotonic_buffer_resource`, which will clear up all
the strings when it is destructed.  We also do the same with
`std::stringstream`.

Change the parameter of `ManifestReader` to take a `std::string_view` to
accept the new values we want to pass it.

A tiny bit of cleaning up of how we use `std::ostringstream` now we have
C++20 and the `view()` method, which saves us some string copies.  We
need to make sure to append a null character to these streams so that
they remain null terminated for ninja's lexer.