Change `OutputText` to use a `std::pmr::list`
To prepare for more complex changes in `trimutil.cpp`, we change the
underlying data structure for `OutputText` to a `std::pmr::list`.  This
will give us O(1) insertion when we need to make changes to the output.
Previously we could only overwrite pieces of the `std::vector`, so we
had to make sure that when we broke up the initial input, we had enough
space for any future changes.

By using a monotonic allocator, we should avoid a lot of the performance
issues going from a `std::vector` to `std::list` as most of the nodes
will be still be contiguous in memory.